social_mcp_server/google-ads-mcp/requirements.txt
# Core
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud
google-cloud-run==0.9.0
google-cloud-pubsub==2.13.11
google-cloud-firestore==2.11.0
google-cloud-bigquery==3.9.0
google-cloud-secret-manager==2.16.1
google-auth-oauthlib==1.0.0
google-ads==19.0.0

# Async
httpx==0.23.3
anyio==3.6.2

# Utils
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3

# Development
autopep8==2.0.2
pylint==2.17.2
pytest==7.3.1
pytest-cov==4.0.0
pytest-asyncio==0.21.0
